fix: label first app as ToDo list
